requirements.txt
flask==3.0.0
opencv-python==4.8.1.78
numpy==1.26.2
pillow==10.1.0
scipy==1.11.4
scikit-image>=0.19.0
